# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Student Management API",
    "description": "API для управління навчальним процесом"
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Введіть токен у форматі: Bearer <Ваш_токен>"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/teachers/add": {
      "post": {
        "tags": ["Teachers"],
        "summary": "Додати вчителя",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Teacher" }
          }
        ],
        "responses": {
          "201": { "description": "Вчителя додано" },
          "400": { "description": "Помилка додавання вчителя" }
        }
      }
    },
    "/teachers/{id}": {
      "delete": {
        "tags": ["Teachers"],
        "summary": "Видалити вчителя",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Вчителя видалено" },
          "404": { "description": "Вчителя не знайдено" }
        }
      },
      "get": {
        "tags": ["Teachers"],
        "summary": "Отримати вчителя за ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Дані вчителя" }, 
          "404": { "description": "Вчителя не знайдено" }
        }
      },
      "put": {
        "tags": ["Teachers"],
        "summary": "Оновити дані вчителя",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/TeacherUpdate" }
          }
        ],
        "responses": {
          "200": { "description": "Дані вчителя оновлено" },
          "404": { "description": "Вчителя не знайдено" }
        }
      }
    },
    "/teachers": {
      "get": {
        "tags": ["Teachers"],
        "summary": "Отримати список усіх вчителів",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Список вчителів" }
        }
      }
    },
    "/teachers/student": {
      "get": {
        "tags": ["Teachers"],
        "summary": "Отримати список вчителів авторизованого учня",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Список вчителів",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Teacher" }
            }
          },
          "404": { "description": "Студента не знайдено" },
          "500": { "description": "Помилка сервера" }
        }
      }
    },
    "/subjects/add": {
      "post": {
        "tags": ["Subjects"],
        "summary": "Додати предмет до студента",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Subject" }
          }
        ],
        "responses": {
          "201": { "description": "Предмет додано" },
          "400": { "description": "Помилка додавання предмета" },
          "404": { "description": "Студента не знайдено" }
        }
      }
    },
    "/subjects/{id}": {
      "delete": {
        "tags": ["Subjects"],
        "summary": "Видалити предмет у студента",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "subjectId", "in": "path", "required": true, "type": "string" },
          { "name": "studentId", "in": "query", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "Предмет видалено" },
          "403": { "description": "Немає доступу до цього предмета" },
          "404": { "description": "Предмет не знайдено" }
        }
      },
      "get": {
        "tags": ["Subjects"],
        "summary": "Отримати предмет за ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Дані предмета" },
          "404": { "description": "Предмет не знайдено" }
        }
      },
      "put": {
        "tags": ["Subjects"],
        "summary": "Оновити предмет",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/SubjectUpdate" }
          }
        ],
        "responses": {
          "200": { "description": "Предмет оновлено" },
          "403": { "description": "Немає доступу до цього предмета" },
          "404": { "description": "Предмет не знайдено" }
        }
      }
    },
    "/subjects": {
      "get": {
        "tags": ["Subjects"],
        "summary": "Отримати список предметів авторизованого студента",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Список предметів" },
          "404": { "description": "Учня не знайдено" }
        }
      }
    },
    "/subjects/student/{studentId}": {
      "get": {
        "tags": ["Subjects"],
        "summary": "Отримати список предметів конкретного студента",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "studentId", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "Список предметів" },
          "404": { "description": "Студента не знайдено" }
        }
      }
    },
...
  },
  "definitions": {
...
    "Teacher": {
      "type": "object",
      "properties": {
        "surname": { "type": "string" },
        "name": { "type": "string" },
        "fatherName": {"type": "string"}
      }
    },
    "TeacherUpdate": {
      "type": "object",
      "properties": {
        "surname": { "type": "string" },
        "name": { "type": "string" },
        "fatherName": {"type": "string"}
      }
    },
    "Subject": {
      "type": "object",
      "properties": {
        "studentId": { "type": "string" },
        "name": { "type": "string" },
        "teacherId": { "type": "string" },
        "hoursPerWeek": { "type": "number" }
      }
    },
    "SubjectUpdate": {
      "type": "object",
      "properties": {
        "name": { "type": "string" },
        "teacherId": { "type": "string" },
        "hoursPerWeek": { "type": "number" }
      }
    },
...
  }
}